requirements.txt
annotated-types==0.7.0
anyio==4.4.0
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
attrs==18.2.0
certifi==2024.7.4
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.8
cssselect2==0.7.0
distlib==0.3.8
distro==1.9.0
Django==5.0.6
django-cors-headers==4.4.0
djangorestframework==3.15.2
filelock==3.15.4
gunicorn==22.0.0
h11==0.14.0
html5lib==1.1
httpcore==1.0.5
httpx==0.27.0
idna==3.7
lxml==5.2.2
markdown2==2.4.13
openai==1.35.10
oscrypto==1.3.0
packaging==24.1
pdflatex==0.1.3
pillow==10.4.0
pipenv==2024.0.1
platformdirs==4.2.2
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pyHanko==0.25.0
pyhanko-certvalidator==0.26.3
pypandoc==1.13
pypandoc_binary==1.13
pypdf==4.2.0
pypng==0.20220715.0
python-bidi==0.4.2
python-dotenv==1.0.1
PyYAML==6.0.1
qrcode==7.4.2
reportlab==4.2.2
requests==2.32.3
setuptools==70.1.1
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.0
svglib==1.5.1
tinycss2==1.3.0
tqdm==4.66.4
typing_extensions==4.12.2
tzlocal==5.2
uritools==4.0.3
urllib3==2.2.2
virtualenv==20.26.3
webencodings==0.5.1
whitenoise==6.7.0
xhtml2pdf==0.2.16
